refactor(settings/recharge): 提取枚举和工具函数到独立文件
